feat/add-ingredient: update ui and add new ingredient
