--- st
@@ -1,51 +1,0 @@
-{
-"name": "Pan-fried salmon with asparagus",
-"userId": 1,
-"video": "https://youtu.be/AvpOyIXs7ns?si=B-61_BTvfI6e_Ops",
-"content": "Pan-fried salmon with asparagus is a suitable choice for busy days when there is not much time to cook. It is both quick and ensures the necessary nutrition, is it worth considering?",
-"tutorial": "<h2>Steps to Prepare:</h2> <h3>Preparing the blood cockles:</h3> <p>After cleaning the blood cockles to remove any mud or debris, soak them in water with a pinch of salt and a few slices of chili to further cleanse. Soak for about 1-2 hours until the cockles release any sand, then rinse thoroughly.</p> <h3>Preparing other ingredients:</h3> <p>Crush and mince the garlic, lemongrass, ginger, and chili for fragrance. Clean and dry the Vietnamese coriander. Lightly crush the roasted peanuts.</p> <h3>Making the tamarind sauce:</h3> <ul> <li>180 ml tamarind sauce (~18 tablespoons)</li> <li>2 tablespoons chili sauce</li> <li>3 tablespoons sugar</li> <li>½ teaspoon MSG</li> <li>1 tablespoon oyster sauce</li> <li>1½ tablespoons fish sauce</li> <li>1 tablespoon annatto oil</li> </ul> <h3>Cooking the stir-fried blood cockles:</h3> <ol> <li>Heat oil and margarine in a pan. When the margarine melts, add the minced garlic, ginger, lemongrass, and chili, and stir-fry until fragrant.</li> <li>Add the tamarind sauce mixture into the pan and cook over medium-low heat.</li> <li>When the tamarind sauce begins to boil, add the blood cockles. Increase the heat and cover the pan to cook the cockles quickly.</li> <li>Stir occasionally to cook the cockles evenly. After 8-10 minutes, the cockles will start to open, indicating they’re cooked.</li> <li>If you prefer a thicker sauce, mix 1 tablespoon of cornstarch with 2 tablespoons of water and gradually add to the sauce. For a thinner sauce suitable for dipping bread, omit the cornstarch.</li> <li>In the final step, add 3 tablespoons of coconut milk to give the dish a creamy richness. Stir well and then turn off the heat.</li> </ol>",
-"imageUrl": "https://ngonaz.com/wp-content/uploads/2021/08/ca-hoi-ap-chao-mang-tay-1.jpg",
-"occasionId": 2,
-"ingredients": [
-{
-"ingredientId": 66,
-"unit": "gram",
-"quantity": 450
-},
-{
-"ingredientId": 83,
-"unit": "gram",
-"quantity": 200
-},
-{
-"ingredientId": 82,
-"unit": "optional",
-"quantity": 1
-},
-{
-"ingredientId": 23,
-"unit": "optional",
-"quantity": 1
-},
-{
-"ingredientId": 3,
-"unit": "optional",
-"quantity": 2
-},
-{
-"ingredientId": 13,
-"unit": "optional",
-"quantity": 1
-},
-{
-"ingredientId": 24,
-"unit": "optional",
-"quantity": 1
-},
-{
-"ingredientId": 16,
-"unit": "optional",
-"quantity": 0
-}
-]
-}
